# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RoVerse Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /planning/boards:
    get:
      operationId: listBoards
      tags: [planning]
      summary: List all planning boards
      responses:
        "200":
          description: Boards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Board"

    post:
      operationId: createBoard
      tags: [planning]
      summary: Create a new board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Created board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
  /planning/boards/{id}:
    get:
      operationId: getBoard
      tags: [planning]
      summary: Get a board with its columns and tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Board with columns and tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardDetail"

    patch:
      operationId: updateBoard
      tags: [planning]
      summary: Update a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"

    delete:
      operationId: deleteBoard
      tags: [planning]
      summary: Delete a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planning/boards/{boardId}/columns:
    post:
      operationId: createBoardColumn
      tags: [planning]
      summary: Add a column to a board
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColumnInput"
      responses:
        "201":
          description: Created column
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardColumn"
  /planning/columns/{id}:
    patch:
      operationId: updateBoardColumn
      tags: [planning]
      summary: Rename or reorder a column
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColumnUpdate"
      responses:
        "200":
          description: Updated column
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardColumn"

    delete:
      operationId: deleteBoardColumn
      tags: [planning]
      summary: Delete a column and its tasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planning/tasks:
    get:
      operationId: listTasks
      tags: [planning]
      summary: List tasks, optionally filtered by board
      parameters:
        - name: boardId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [planning]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /planning/tasks/{id}:
    get:
      operationId: getTask
      tags: [planning]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    patch:
      operationId: updateTask
      tags: [planning]
      summary: Update a task (move column, assign, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      operationId: deleteTask
      tags: [planning]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /planning/events:
    get:
      operationId: listEvents
      tags: [planning]
      summary: List calendar events
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"

    post:
      operationId: createEvent
      tags: [planning]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /planning/events/{id}:
    patch:
      operationId: updateEvent
      tags: [planning]
      summary: Update a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"

    delete:
      operationId: deleteEvent
      tags: [planning]
      summary: Delete a calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Board:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BoardInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
        teamId:
          type: integer
    BoardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
    BoardColumn:
      type: object
      required: [id, boardId, name, position]
      properties:
        id:
          type: integer
        boardId:
          type: integer
        name:
          type: string
        position:
          type: integer
    BoardDetail:
      type: object
      required: [id, name, columns, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnWithTasks"
        createdAt:
          type: string
          format: date-time
    ColumnWithTasks:
      type: object
      required: [id, boardId, name, position, tasks]
      properties:
        id:
          type: integer
        boardId:
          type: integer
        name:
          type: string
        position:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    Task:
      type: object
      required: [id, boardId, columnId, title, position, createdAt]
      properties:
        id:
          type: integer
        boardId:
          type: integer
        columnId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeUsername:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        dueDate:
          type: ["string", "null"]
          format: date
        position:
          type: integer
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    TaskInput:
      type: object
      required: [boardId, columnId, title]
      properties:
        boardId:
          type: integer
        columnId:
          type: integer
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
        tags:
          type: array
          items:
            type: string
    TaskUpdate:
      type: object
      properties:
        columnId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        dueDate:
          type: ["string", "null"]
          format: date
        position:
          type: integer
        tags:
          type: array
          items:
            type: string
    CalendarEvent:
      type: object
      required: [id, title, startDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        color:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        allDay:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CalendarEventInput:
      type: object
      required: [title, startDate]
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        color:
          type: string
        assigneeId:
          type: integer
        allDay:
          type: boolean
    CalendarEventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        color:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        allDay:
          type: boolean
    ColumnInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        position:
          type: integer
    ColumnUpdate:
      type: object
      properties:
        name:
          type: string
        position:
          type: integer
...